Allow gopsutil docker to build on non-linux boxes by stubbing out non-linux libraries
